+++ core/src/main/java/org/lemurproject/galago/core/retrieval/iterator/BiL2ScoringIterator.java
@@ -41,7 +41,7 @@ public class BiL2ScoringIterator extends TransformIterator implements ScoreItera
   }
 
   @Override
-  public void syncTo(int identifier) throws IOException {
+  public void syncTo(long identifier) throws IOException {
     super.syncTo(identifier);
     lengths.syncTo(identifier);
   }
