+++ core/src/main/java/org/lemurproject/galago/core/retrieval/iterator/InL2ScoringIterator.java
@@ -45,7 +45,7 @@ public class InL2ScoringIterator extends TransformIterator implements ScoreItera
   }
 
   @Override
-  public void syncTo(int identifier) throws IOException {
+  public void syncTo(long identifier) throws IOException {
     super.syncTo(identifier);
     lengths.syncTo(identifier);
   }
