+++ core/src/test/java/org/lemurproject/galago/core/retrieval/query/NodeTypeTest.java
@@ -82,12 +82,12 @@ public class NodeTypeTest extends TestCase {
     }
 
     @Override
-    public void movePast(int identifier) throws IOException {
+    public void movePast(long identifier) throws IOException {
       throw new UnsupportedOperationException("Not supported yet.");
     }
 
     @Override
-    public void syncTo(int identifier) throws IOException {
+    public void syncTo(long identifier) throws IOException {
       throw new UnsupportedOperationException("Not supported yet.");
     }
